Allow displayed diversity data to be downloaded
We frequently get requests to share the data displayed in the
diversity panel. Here we address this by allowing download of
the data behind the graph. This is the simplest implementation,
as to download all data together ({events,entropy}U{nt,aa})
would require recalculations within the download logic as these
data are not stored by auspice when not being displayed.